+++ R/transform.R
@@ -158,6 +158,17 @@
 #'       are only interested in the fuzzy input graph (e.g. for clustering),
 #'       setting \code{n_epochs = 0} will avoid any further sparsifying.
 #'   }
+#' @param seed Integer seed to use to initialize the random number generator
+#'   state. Combined with \code{n_sgd_threads = 1} or \code{batch = TRUE}, this
+#'   should give consistent output across multiple runs on a given installation.
+#'   Setting this value is equivalent to calling \code{\link[base]{set.seed}},
+#'   but it may be more convenient in some situations than having to call a
+#'   separate function. The default is to not set a seed, in which case this
+#'   function uses the behavior specified by the supplied \code{model}: If the
+#'   model specifies a seed, then the model seed will be used to seed then
+#'   random number generator, and results will still be consistent (if
+#'   \code{n_sgd_threads = 1}). If you want to force the seed to not be set,
+#'   even if it is set in \code{model}, set \code{seed = FALSE}.
 #' @return A matrix of coordinates for \code{X} transformed into the space
 #'   of the \code{model}, or if \code{ret_extra} is specified, a list
 #'   containing:
@@ -201,7 +212,8 @@ umap_transform <- function(X = NULL, model = NULL,
                            learning_rate = NULL,
                            opt_args = NULL,
                            epoch_callback = NULL,
-                           ret_extra = NULL
+                           ret_extra = NULL,
+                           seed = NULL
                            ) {
   if (is.null(n_threads)) {
     n_threads <- default_num_threads()
@@ -238,6 +250,30 @@ umap_transform <- function(X = NULL, model = NULL,
     else {
       n_epochs <- max(2, round(n_epochs / 3))
     }
+  }
+
+  # Handle setting the random number seed internally:
+  # 1. If the user specifies seed = FALSE, definitely don't set the seed, even
+  # if the model has a seed.
+  # 2. If the user specifies seed = integer, then use that seed, even if the
+  # model has a seed.
+  # 3. If the user does not specify a seed, then use the model seed, if it
+  # exists. Otherwise don't set a seed. Also use this code path if the user
+  # sets seed = TRUE
+  if (is.logical(seed) && !seed) {
+    # do nothing
+  }
+  # handle the seed = TRUE case in this clause too
+  else if (is.logical(seed) || is.null(seed)) {
+    if (!is.null(model$seed)) {
+      tsmessage("Setting model random seed ", model$seed)
+      set.seed(model$seed)
+    }
+    # otherwise no model seed, so do nothing
+  }
+  else {
+    tsmessage("Setting random seed ", seed)
+    set.seed(seed)
   }
 
   if (is.null(search_k)) {